+++ ChangeLog.md
@@ -2,10 +2,12 @@
 
 v1.0.1
 
+- Updated MediaPlayer.Avalonia volume design
 - Added code documentation
 - Moved Source and Title properties from implementations to PlayerHostBase
 - Updated online documentation
 - Implemented IDisposable for MediaPlayer.Wpf.Mpv
+- MediaPlayer.Avalonia.Bass, fixed tempo shift calculation error
 - MediaPlayer.Avalonia.Bass, added EffectsAntiAlias and EffectsAntiAliasLength to customize output
 - MediaPlayer.Avalonia.Bass, added EffectsQuick to reduce CPU usage
 - MediaPlayer.Avalonia.Bass, added EffectsSampleRateConversion and EffectsFloat to improve quality
@@ -15,5 +17,4 @@ v1.0.1
 - BASS errors now throw BassException instead of InvalidOperationException
 
 TODO:
-- Updated MediaPlayer.Wpf.Mpv to support `mpv-2.dll`
-- Updated MediaPlayer.Avalonia volume design
+- Update MediaPlayer.Wpf.Mpv to support `mpv-2.dll`